[REF] share child articles
With this commit, we add the ability to manage tokens and share
child articles with the same token.

An article shared with a token will display the following
- Sidebar of child articles if shared
- Ability to search for available articles by name
- Links in the article will be replaced with the token link

We can manage articles using the Share button in the backend view, or
we can access the Articles Shared Tokens menu under
Knowledge -> Configuration.